pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fiscore-backend"
version = "0.1.0"
description = "FiScore backend scaffold for API, ingestion worker, and source integrations."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.13.0,<5.0.0",
  "fastapi>=0.116.0,<1.0.0",
  "google-cloud-storage>=3.1.0,<4.0.0",
  "httpx>=0.28.0,<1.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0,<9.0.0",
  "ruff>=0.12.0,<1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/fiscore_backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]
